// src/commons/types/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    Date: any;
    Upload: any;
};
// ...
export type IMutation = {
    __typename?: "Mutation";
    createBoard?: Maybe<IReturn>;
    createProduct?: Maybe<IReturn>;
    createProfile?: Maybe<IReturn>;
    deleteBoard?: Maybe<IReturn>;
    deleteProduct?: Maybe<IReturn>;
    deleteProfile?: Maybe<IReturn>;
    updateBoard?: Maybe<IReturn>;
    updateProduct?: Maybe<IReturn>;
    updateProfile?: Maybe<IReturn>;
};

export type IMutationCreateBoardArgs = {
    contents?: InputMaybe<Scalars["String"]>;
    title?: InputMaybe<Scalars["String"]>;
    writer?: InputMaybe<Scalars["String"]>;
};
// ...
export type IReturn = {
    __typename?: "Return";
    _id?: Maybe<Scalars["String"]>;
    message?: Maybe<Scalars["String"]>;
    number?: Maybe<Scalars["Int"]>;
};